Starting edition on initialScreen
