+++ LetsTogether/LetsTogether/Scenes/ViewModel/InitialScreenViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  InitialScreenViewModel.swift
+//  LetsTogether
+//
+//  Created by José Mateus Azevedo on 16/09/21.
+//
+
+import Foundation
